Fix race between TLSCredProcessor and ThriftServer
Summary:
TLSCredProcessor runs a background thread that has references back into ThriftServer and to another component (sharedSSLContext_) that ThriftServer may modify.

The lifetime of `sharedSSLContext_` prior to this diff does not strictly exceed the lifetime of `TLSCredProcessor`. `sharedSSLContext_` may be destroyed (by `stopAcceptingAndJoinOutstandingRequests()`) while the polling thread is still active.

In reality, `TLSCredProcessor` really needs to be scoped to just the function call in `ThriftServer::serve()`. However, this is made difficult to to the fact that `TLSCredProcessor` is exposed indirectly through public accessors.

This diff destroys the tlscredprocessor in the `ThriftServer::serve()` cleanup routine so that it occurs prior to any modifications to `sharedSSLContext_`.

Reviewed By: andriigrynenko

Differential Revision: D25041860

fbshipit-source-id: b61ba248a0751ad9e9f689b5a35df44b83660abd