rust: fix skipping of unexpected empty sequences
Summary:
If a thrift structure is changed to add a new sequence (list, set, or map), and
the new implementation includes an empty sequence, then a parser for an old
version will fail to skip the new sequence when parsing the binary encodings of
that stucture.

This is because, when skipping the sequence, we ignore the fact that
`read_list_begin` returns a length of 0, and go ahead and try to parse the
first value.  Because the next bytes in the binary encodings are the following
item, we'll try to parse that item as the type of the list element, and fail.

The fix is to not parse any values if we are told the list is empty.

Reviewed By: yfeldblum

Differential Revision: D25088231

fbshipit-source-id: b1725b73895999d32a258b92fda24a90c2748366