Filter connection events lazily based on the first incoming request.
Summary:
We used to log connection events when we know the security type used for each connection, now in order to get rid of some noise before 100% rollout to rocket we postpone this only when the first request comes in for headers.

And for rocket we only log during setup frame.

(Note: this ignores all push blocking failures!)

Reviewed By: andriigrynenko

Differential Revision: D24201735

fbshipit-source-id: 5491f408f7f9b2cb87032aaec4e329bca005a5d6