Add access to generated ThriftTypeInfo
Summary: This fully decouples the ThriftTypeInfo from the "any" option. With this decoupling, we can no longer use the absence of the "any" option to avoid a circular dependency. To avoid the dependency issue, the generation of the ThriftTypeInfo has been moved out of the code gen, and now happens lazily va getGeneratedThriftTypeInfo.

Reviewed By: iahs

Differential Revision: D25069795

fbshipit-source-id: 1ce64a75c2e66a74a0c24ab308c81b31ba6ef03f