Read connection metadata into ConnectionLoggingContext for rocket and header.
Summary:
On request setup, there's metadata attached and we want to store them somewhere with the connection lifetime for subsequent logging calls.

Therefore we add a new class ConnectionLoggingContext added to store the transport type, the client host id and the client agent name. For THRIFT_CONNECTION_EVENT() we add additional arguments to record logging context as well.

Reviewed By: andriigrynenko

Differential Revision: D25226215

fbshipit-source-id: 4b659da2efe8f4d318ef8ce09b52a13822c31c76