add connection context to preprocess handler
Summary:
the purpose of `preprocess` callback is to allow for scenarios like ratelimiting, which is often based on the identity of the client.
Because with secure connections we do authenticate clients, connection contains information about the client that can be used in such scenarios. So lets exposed connection to preprocess callbacks.

Reviewed By: zhxchen17

Differential Revision: D25230617

fbshipit-source-id: 20d75ed78ba696bb4b7931c92a754cc57ec58c55