allowing coexistence of functions with/without stack arguments returning same type
Summary:
Currently when you have a function with stack_arguments enabled and another not enabled, and they both return the same type, the compile will fail as two `cdef class Promise_<ReturnType>` will be defined in the same module, which causes a name conflict.

Fixing this by adding "__sa" to the promise name when stack_arguments are enabled.

Reviewed By: yfeldblum

Differential Revision: D25134851

fbshipit-source-id: 75d3ce2df2aab418d79e50fc492df77abdd935b7