Knobs for limiting memory usage for rocket transport layer
Summary: In the case of high connection fan-in and large request, it is possible leading to transport layer allocate too much memory. This diff provides a knob to control how much memory transport layer would allocate memory for the requests, and will drop the connection if certain criteria is met.

Reviewed By: iahs

Differential Revision: D24958410

fbshipit-source-id: f78c5a1ea98c1636fea3d768d061272ec180a78c