Fix null reference accessing from DebugStub getters
Summary: In Thrift server requests registry, DebugStub can be mapped to both active requests and finished requests. For finished requests since the request objects are gone, debug stub no longer references to them so that we have null pointer for `req_`, and in this case when we are returning the request object from stub, we should return null pointer instead of null reference because the later one is not well defined in cpp standard.

Differential Revision: D25540104

fbshipit-source-id: 210d9749ba63fea9edc92293c58182ea7f896534