keep asynchronous tasks on the intended thread pool - part II (codegen)
Summary:
Today, a continuation of coro or semifuture application handlers will always execute on NORMAL priority thread pool. This is because the effective executor for the task is the ThreadManager::Impl and its add() implementation does not preserver tasks' priority (which we map into thread pool).

In this change, Generated code is routing either coro or semifuture onto the corresponding keepalive, which is preserved across tasks that do not explicitly specify a target executor.

Reviewed By: zhxchen17

Differential Revision: D25140498

fbshipit-source-id: 34c9d1d7df0e589a2afa4d6ae830b4aef936bd26