Simplify handling of chunkBufferSize = 0
Summary:
We currently send a 0 credit message and then immediately a 1 credit message on every loop iteration. We can instead reduce this case to the size = 1 case before the loop and remove the extra check.

The purpose of supporting 0 here is to allow deferring payload receipt until the subscription begins; once it has there is no need to treat this differently from a size of 1.

Reviewed By: rhodo

Differential Revision: D25537142

fbshipit-source-id: 48245e1b4d8db865f1aa2fe22bc445346f50c43d