Allow "oneway" identifier
Summary:
In prep for adding "idempotent" and "readonly" as additional function modifiers, without breaking compatibility with existing idents. "idempotent","readonly", and "oneway" are mutually exclusive, so should ideally behave the same way.

This change makes it so keywords can be explicitly allowed as idents (as long as it doesn't create ambiguity). I've updated 'sink' to use this simplified method. This makes it so 'sink' can be used as a function name or function argument names (previously disallowed). I could continue to disallow 'sink' (and 'oneway') as a function name, but it doesn't look like this creates an ambiguity, and adding the restriction would make the parser (slightly) more complex.

Reviewed By: vitaut

Differential Revision: D25543390

fbshipit-source-id: 01ce0227cd452b02b686300a8adee2fabbc5c15f