Java: Disable generation of `fieldIsSet*` series of methods by default
Summary:
A generated Java class (from the java.swift codegen) generates a set of methods
prefixed by `fieldIsSet`, the purpose is to check if a field has been set or if
the default value has been applied by the deserializer.
It's been added for easing the calisto migration away from `javadeprecated`
codegen.

With this change, the new default behavior is to NOT generate `fieldIsSet*`
methods, but you can enable this feature by adding
`(java.swift.enable_is_set = "true")` on the Thrift struct (in the IDL).

Differential Revision: D25605109

fbshipit-source-id: 94b3ca0a42f9e48a1675f8905243f80179ad4800