Simplify and fix fromMap_DEPRECATED codegen
Summary: The existing implementation of `frommap_construct` had occasional errors, e.g. pass default 0 instead of null for optional fields, leading to test failures like P155590363 when removing mapconstruct. Also, most of the complexity can be delegated to the constructor while keeping this a thin wrapper with a bunch of HH_FIXME.

Reviewed By: prasad223

Differential Revision: D25631338

fbshipit-source-id: 10da1d32d6b031e9f76d1fd74e89f61e6d30af41