Fix and simplify arithmetic right shift check in compact
Summary:
The check for signed right shift being arithmetic in the compact protocol implementation doesn't do anything because there is no `configure` script anymore and its results are hardcoded. Replace the configure-based check with a simple static assert that actually works.

As a nice side effect this removes one dependency.

Reviewed By: iahs

Differential Revision: D25783207

fbshipit-source-id: 7ad4dc1f4bb438080e1ed97eeabb43a7136a4a91