Add hack.adapter annotation support to the hack generator
Summary:
Replace the field type with Adapter::THackType everywhere. Also use the adapter to convert the default value for non-optional fields.

The most tricky part is typedefs, where we want to respect the annotation regardless of whether it's inside or outside the typedef (also applies to struct names). For this I had to move around some of the `get_true_type` calls to make sure we can check adapters before the typedefs are resolved. Some calls were redundant (downstream code would call it again anyway) and are removed completely.
Several scenarios that are not supported (yet?):
1. map/set keys - previously discussed, would require additional changes to the serialization code
2. shapes and json compiler options - no major technical blockers, just doesn't seem worth it for now (both have low adoption and appear on the way out - correct me if I'm wrong)

Reviewed By: andrii-korotkov

Differential Revision: D25626221

fbshipit-source-id: 14b3f1495233d879183996e9dcd41ee0e8dfb5ad