Get hh to validate IThriftAdapter::TThriftType
Summary:
There is a potential problem in that a developer might specify the adapter with the wrong type for the field it's used on. Eg here the field type is `Bar`, but the adapter actually accepts `Foo`:
```
class MyAdapter implements IThriftAdapter {
  const type TThriftType = Foo;
  // ...
}

struct MyStruct {
  1: Bar (hack.adapter = "MyAdapter") bar;
}
```

In such a scenario there would be no enforcement at the type checking time since adapters are only used by untyped code, but at runtime things will break in unpredictable ways (hello php...).

To get hh catch such errors, we will generate a dummy function that is never called but only exists for the typechecker to complain if the types are wrong.

Reviewed By: andrii-korotkov

Differential Revision: D25675514

fbshipit-source-id: 5f0a74bde31befbdbc5cc9c054151ce17b2e8c2b