Generated RpcServerHandler
Summary:
Generates an `RpcServerHandler` from an IDL instead of using reflection, and annotations. Still provides support for mixing together services by implementing a method called `getMethodMap`. This provides a map of methods names. A class called `CompositeRpcServerHandler` has a constructor that takes one or more `RpcServerHandlers`. It uses the `getMethodMap` to create a mapping between requests and service methods.

This will be used to support streaming eventually.

Reviewed By: avalonalex

Differential Revision: D25835312

fbshipit-source-id: c8c4a75f774f8c7addc4c349e8aa0f2af39a6170