Add `idempotent` and `readonly` keywords.
Summary:
This diff just adds the keywords and updates the AST. It doesn't actually change the code gen.

idempotent -> can be retried immediately (even if the server may have processed the request)
readonly -> can be retried any time (even if the previous request succeed a week ago). Note that for HTTP this is know as "safe"

Differential Revision: D25288166

fbshipit-source-id: 070e7cd35a450e740b8512e3127f4654ad4e605b