add container cache util
Summary:
if a struct has a list field and the list was used to iterate for multiple times, new Python objects are created at each iteration, as a list field is not a true python list but a Cython `List__xxx` type which wraps a c++ container object (e.g. `std::vector`), and every iteration will need to create Python objects from the c++ container on the fly. The generated Python objects are not kept/cached.

We may cache all the generated objects but that may incur a perf/memory cost if most of the list are iterated only once. Before we make a decision to cache or not cache the containers by default, it's good to have an explicit cache util function for use cases that are known to iterate containers for multiple times.

Reviewed By: junzh0u

Differential Revision: D25895579

fbshipit-source-id: 6c52b0aa2726b6f64e861df36bc69f7e25ac8c4b