Improve build times by moving Thrift data from headers to source files
Summary:
Move Thrift-generated data from `*_data.h` to `*_data.cpp` to reduce build times.

This removes `constexpr`-ness of data which nobody relies on at the moment. If we find a use case in the future we can provide `constexpr` access to data via an additional header limiting the compile-time overhead just to use cases that need it. This won't affect inherently dynamic use cases such as py3.

Reviewed By: Mizuchi

Differential Revision: D26057730

fbshipit-source-id: 8386c8205796f55ca30cb338590c5d6358a2d9cb