Add --allow-experimental-features flag
Summary:
...to gate experimental IDL features.

As this flag protects IDL features, it is target language agnostic. The flag is represented by set of strings to indicate which features should be allowed. Features can be enabled individually by name (`--allow_experimental idempotent,readonly`) or all at once (`--allow_experimental all`, likely should only be used by tests).

To guard a feature, the helper function 'require_experimental_feature' is provided, which checks if the experimental feature is enabled, and returns a fatal error if not.

When a feature is ready for general availability, the require_experimental_feature check can be removed without breaking any references to the feature in a allow_experimental flags (though these references would ideally be cleaned up).

In subsequent diff, I will add support to the Thrift Buck build rules to enable features.

Reviewed By: vitaut

Differential Revision: D25919464

fbshipit-source-id: d08b448639b63a75d2ce4acb16ae27a8ad222d62