Remove "set_cpu_worker_threads" API and update related callsites
Summary: In Python, there is only one server thread doing the work. Thus `set_cpu_worker_threads` is misleading and should be removed. In the past the only work that was happening on these threads was request payload deserialization and response payload serialization. Subsequent D26022206 is changing that to execute ALL work on the single python thread that is running the event loop.

Reviewed By: nanshu

Differential Revision: D26131782

fbshipit-source-id: c2acb55a9af624f12407d8698e4a2f30eae00c2a