Start to send structured ClientConnectionMetadata instead of host and agent headers from client.
Summary:
To better expose client data from Cpp2ConnContext, and at the same time make it easier to update the data structure we pass over the wire, we can group client metadata in a single Thrift struct and store it in Cpp2ConnContext.

This diff is the first part: Adding ClientConnectionMetadata and sending it over the wire from the client side.
On the server side we will extract it from header/RpcMetadata but not store it, which will be left to the next diff.

Reviewed By: andriigrynenko

Differential Revision: D25964631

fbshipit-source-id: f5f1dc5db3f43effab238542a175238d161ab185