introduces FixedSizeString type.
Summary:
FixedSizeString is sub-type of string that has a fixed size specified in the IDL schema. The size is agreed by both the encoder and decoder so that the frozen data doesn't have to contain it, making the frozen data smaller than the raw string type.

Note FixedSizeString field can only be unset if it's optional. An exception will be thrown for unqualified fields if unset.

Reviewed By: ddrcoder

Differential Revision: D26166320

fbshipit-source-id: e6fd9916db9890712b70a82f78f03af73d5ca28f