thrift: Stochastic fair queue thread manager (SFQTM)
Summary:
## Stochastic fair queue thread manager (SFQTM)
This TM variant aims to address fairness concerns in multi-tenant systems by
fronting the CPU executor' UPSTREAM queue with a configurable number of
sub-queues used to isolate different request "tenants". Usage of the SFQTM
relies on users to identify different request tenants via including the data
in the execution scope provided to getKeepAlive() calls for UPSTREAM
sources.

### Fairness in concurrency
Fairness in the SFQTM context means that each tenant is allowed the same
number of tasks to reside on the internal executor queue. As an example,
let's say we have 2 tenants (A and B). When the internal executor's threads
are busy running tasks and begins queuing tasks, its queue will contain at
most a single task from A and a single task from B. As tasks from each tenant
are scheduled on a thread, a new task for that tenant is pulled from the
tenant's queue and placed in the internal executor's queue.

### Stochastic fairness
For practical reasons related to memory utilization and performance, multiple
tenants can potentially map to a single queue. When instantiating the
SFQTM, we allocate a fixed number of queues and hash the tenant ID onto this
fixed set of queues. This means that there is potential for multiple tenants
to map onto the same queue, resulting in a shared request concurrency
entitlement between them and unfairness. We mitigate this via perturbing the
hash function periodically via 'perturb'. This ensures that even if there is
a collision between two tenants, the condition will not be persistent. This
perturbation may result in request reordering for a tenant under heavy load.

### Notes
* Request concurrency is limited per-tenant (or priority). Therefore,
  'maxOutstanding' must be set such that there is efficient/acceptable use of
  the executor in the case of workloads that consist of a single tenant.

* One may disable perturbing of the tenant hash by setting the period to
  0 seconds.

* If unspecified in the execution scope, tenant ID is assumed to be
  zero.

Differential Revision: D25842631

fbshipit-source-id: 1de57bc81b8bc9fca9b915a74d6d9d9f83b5454c