Add size target buffering policy to ClientBufferedStream
Summary:
Adds API to ClientBufferedStream that attempts to maintain constant memory usage in buffer instead of constant payload count.
chunkBufferSize is still used as the initial number of credits to send but is ignored afterwards. Since this only affects the buffering behavior after subscription it seemed like this belongs on ClientBufferedStream instead of on RpcOptions, but we could also expand chunkBufferSize into a streamBufferingPolicy struct.
Assumes outstanding payloads will come back with max size, which is a conservative estimate that converges to a safe state quickly for normal payload size distribution but will substantially reduce the effective buffer size when variance is large. Any such empirical estimate is subject to a class of pathological streams (e.g. payloads that increase monotonically in size for longer than the chunkBufferSize); we could add a knob for setting a prior or limit on the size, but I'm skeptical anyone will want or know how to tune that. Absent that, more aggressive strategies like maintaining a rolling average payload size are likely to go over target often enough to be ineffective at maintaining high memory utilization.

Reviewed By: clairezhang-fb

Differential Revision: D25562513

fbshipit-source-id: 6751bd15b3c5b5935b40465ec39aa9894e2d5c88