handle default map/set/list values for mutable structs
Summary:
if a thrift struct is mutable, we should probably not use Immutable version of Map/Set/List for default value.

In this diff, if a struct is mutable, we just wrap the value with corresponding `Maps.newHashMap`, `Sets.newHashSet` and `Lists.newArrayList` correspondingly.

This change should only affect mutable structs.

Reviewed By: prakashgayasen

Differential Revision: D26239667

fbshipit-source-id: 07c879c2dcb6145fef572d70bede5434ab019121