Add ServerTracker API
Summary:
Server tracker is an object with a scoped lifetime, an associated key and an associated Thrift server object reference, which can be created to instrument feature uses from thrift servers.

We also unify the previous ServerInstrumentation API into this one, so that we only need to call a free function `apache::thrift::forEachServer()` to iterate over existing Thrift servers, which is implemented based on ServerTracker.

Differential Revision: D22872641

fbshipit-source-id: 36cbca449ab0cbe06f42fbd1779b302f2491f89f