thrift: Add getRequestExecutionScope() to the server interface
Summary:
This patch adds the `getRequestExecutionScope` function to the server interface which is  used to find the appropriate
executor. Services may implement the function if they use a thread manager implementation that will make full use of execution scope (such as the SFQThreadManager).

getRequestPriority() is left in place, but will be removed in a separate patch.

Differential Revision: D26118802

fbshipit-source-id: 7931466ab23abc462ba15ca00eb00002c6489fd2