Generate unique names for optional constants
Summary:
Consider the following thrift code:
  struct PortsRange {
      1: i32 start
      2: optional i32 end
  }
  const list<PortsRange> MYLIST = [
      PortsRange{start=13, end=37},
      PortsRange{start=42, end=43},
  ]
  const map<string, PortsRange> MYMAP = {
      "foo": PortsRange{start=13, end=37},
      "bar": PortsRange{start=42, end=43},
  }
  const set<PortsRange> MYSET = [
      PortsRange{start=13, end=37},
      PortsRange{start=42, end=43},
  ]

The generated Go code is wrong and doesn't compile.

Go needs to generate global variables to hold the const values of optional fields.
The name of the generated variables are created using the type, the name of the container and the variable. Example: `const_lit` + `MY_LIST` + `end`.

The problem is that it doesn't take into account the index or key and we end up with duplicates when we initialize optional fields in a const list/map/set.

With the thrift code above, the generate code looks like this:
  var const_lit_MYLIST_end int32 = 37
  var const_lit_MYLIST_end int32 = 43  // duplicate
  var const_lit_MYMAP_end int32 = 37
  var const_lit_MYMAP_end int32 = 43   // duplicate
  var const_lit_MYSET_end int32 = 37
  var const_lit_MYSET_end int32 = 43   // duplicate
  ...
  MYLIST = []*PortsRange{
    &PortsRange{
        Start: 13,
        End: &const_lit_MYLIST_end,
    },
    &PortsRange{
        Start: 42,
        End: &const_lit_MYLIST_end, // wrong
    },
  }
  ...

Same applies for sets and maps.

This diff addresses the issue of generating unique variable names by appending an index at the end of the container name:
  var const_lit_MYLIST_0_end int32 = 37
  var const_lit_MYLIST_1_end int32 = 43
  var const_lit_MYMAP_0_end int32 = 37
  var const_lit_MYMAP_1_end int32 = 43
  var const_lit_MYSET_0_end int32 = 37
  var const_lit_MYSET_1_end int32 = 43
  ...
  MYLIST = []*PortsRange{
    &PortsRange{
      Start: 13,
      End: &const_lit_MYLIST_0_end,
    },
    &PortsRange{
      Start: 42,
      End: &const_lit_MYLIST_1_end,
    },
  }

Differential Revision: D26312417

fbshipit-source-id: 5a2b1e87ae38e3febe1a4bb1bdeb9369a5ae77ea