reflection name attrs should be idl names
Summary:
In reflection, the name reflection attributes of all things should be compile-time constexpr representations of idl names, not of c++ names.

Applies to struct field names, union field names, and enum value names - these are the things for which reflection honors renaming.

Affects `to_dynamic` and `from_dynamic` - presumed to be a breaking change when used with renamed fields and, in some formats, with renamed enum values.

Reviewed By: iahs

Differential Revision: D26413154

fbshipit-source-id: 80770fb798e8a7aea6364bb7b2d491a61e36880f