add support for Bytes as rust.type for binary fields
Summary:
Allow specification of `rust.type = "Bytes"` for `binary` types in
thrift definitions.  This uses `bytes::Bytes` as the concrete type
for the field.

As a special optimization, if the thrift structure is being deserialized from a
`Bytes` buffer, then the `Bytes` in the deserialized thrift structure will
share data from that `Bytes` buffer.

Reviewed By: mitrandir77

Differential Revision: D26484050

fbshipit-source-id: 3a239819576ead1ddf1ac7d4c6b58d996cc353a5