add default field similar to protobuf
Summary:
Protobuf structs have a default field that you can use when you want something default. It's also static so you don't have to create a default object repeatedly. Instead of a default field, I added a static default() method that returns the value a default object. For instance if you're making a call to an RPC method, and you want to pass in a default struct you can use this to get a default object.

The default method will only appear on immutable structs.

Reviewed By: avalonalex

Differential Revision: D26489423

fbshipit-source-id: 086c6e7d654360829aace3b3f0e789168faf446f