API changes for metadata to enable recursive calling of structs and services
Summary: The way the thrift api is defined is fairly inconvenient for recursive usage, as the child of a struct's fields will only contain a string telling you what type they reference, and one would have to return to the return dictionary to check what type that might be. This adds a python wrapper to the c++ api that handles this implicitly.

Reviewed By: nanshu

Differential Revision: D25878113

fbshipit-source-id: 3541d4a9d1fced4b3e01d58803c28c104b274280