Use Enum.fromInteger(0) as default value for an enum
Summary:
Currently, we use `null` as the default value for an enum.
It is correct when an enum doesn't have an entry for `0`.
However, when an enum has a value for `0`, its default value should be this entry to mimic the behavior of other languages.

E.g.:
```
enum Color {
    GREEN = 1,
    BLUE = 2,
}

enum Taste {
    SALTY = 0,
    SWEET = 1,
}

struct Foo {
    1: Color color,
    2: Taste taste,
}
```

The default value of `Foo.color` must be `null` while the default value for `Foo.taste` must be `SALTY`.
In the past, both `Foo.color` and `Foo.taste` were initialized with `null`.

__Note__: This code doesn't change anything on the wire, we are still sending `0` for an enum. The serialized data is byte to byte identical.

#forcetdhashing

(Note: this ignores all push blocking failures!)

Differential Revision: D24915333

fbshipit-source-id: a61f1e8754e8af14f49b123defbcc08d883040dc