Renamed and reworked t_types
Summary:
Part of the conversion of enum members from UPPER_CASE to snake_case. Cleaning up enum/class definitions.

Note: For now keeping the members as is (ie UPPER_CASE). Will convert as part of a followup diff, also introduced 2 typedefs to make this a smaller diff for the time being, and which will be removed.

Reviewed By: yfeldblum

Differential Revision: D26555757

fbshipit-source-id: efe750fecaab77218f90b568983fe806a8999248