Introduce API for handling compressed SerializedRequests
Summary:
If a rocket request payload is compressed, it currently gets decompressed by the IO worker thread. This change introduces a new API on AsyncProcessor to accept a still compressed SerializedRequest, which can be decompressed at a later stage (i.e. on the CPU worker).

By default, the new API will decompress and forward the result to the old API for compatibility with custom processors that derive directly from AsyncProcessor.

In GeneratedAsyncProcessor, the old API is now marked final and instead forwards to the new one, which is now implemented by all generated processors instead.

In cases where custom processors are derived from generated ones instead of AsyncProcessor, the overrides are adjusted accordingly to the new interface.

For now, SerializedRequests are still decompressed by the server handler and forwarded without compression.

#buildmore
#forcetdhashing
#keepgoing

Differential Revision: D26324349

fbshipit-source-id: be17a14da5edfbe077c3c8e277f969070fe7e884