simplify ResponseChannelRequest. Remove getBuf/extractBuf APIs
Summary:
We do not pass payload as part of the ResponseChannelRequest. This may be used
to be the case long time ago, but nowadays there is `SerializedRequest` that we
pass explicitly to processor.
The only place where payload is still stored on request context is in
`HeaderRequest` which is passed to Cpp2Request and is never directly used
anywhere else.

Reviewed By: sethdelliott

Differential Revision: D26693897

fbshipit-source-id: 4578d3fdb5127eee70c966b41205e89099070f8b