simplify ResponseChannelRequest. Remove getBuf/extractBuf APIs
Summary:
We do not pass payload as part of the ResponseChannelRequest. This may be used
to be the case long time ago, but nowadays there is `SerializedRequest` that we
pass explicitly to processor.
The only place where payload is still stored on request context is in
`HeaderRequest` which is passed to Cpp2Request and is never directly used
anywhere else.

Reviewed By: yfeldblum

Differential Revision: D26711267

fbshipit-source-id: 148b0743a688555d4daf9b172d191fe99b5c61b8