fix rocket upgrade tests
Summary:
* client thread is sending multiple requests on the client async interface, over a channel that is currently bound to an event base running on a different thread. The client hence needs to access event base stored on the channel (in order to transition over to that thread). Notice, however, that the event base is stored on the socket/transport object.
* at the same time, the io thread has probed the server that it does supports rocket, and is resetting the transport on the channel (to nullptr).
* the main thread then cannot access the eventbase

The problem is that the channel infers its attached eventBase from transport, and, once transport is briefly reset so is the event base, it breaks the safety assumptions about reading attached event base from another thread breaks.

The solution is to keep a reference event base in a dedicated field, derive it from transport when transport is set. derive it from transport on attach/detach events, but avoid resetting the event base it when transport is set to nullptr.

Reviewed By: yfeldblum

Differential Revision: D26710007

fbshipit-source-id: a3c6813d693a7e679cdeff00e11d3b859f4e2c00