Add exception qualifiers
Summary:
Like existing qualifiers, these come before the 'introducer' with the form:
  (safe)? (transient|stateful|permanent)? (client|server)? exception MyExcept {...}

This information will be propagated back from the server through rpc response metadata to the client when an error is thrown. The client will use this information to decide if a retry is safe. This information will also be used for generic monitoring.

Reviewed By: vitaut

Differential Revision: D26533617

fbshipit-source-id: d021e08c458ad4f99442fc24ae8ab96c65b3b4db