Java: Add code generation for Interactions
Summary:
Generate code for Thrift `Interactions`, but without implementations since we need to wait for Rocket to be available.

I tried to mimic the C++ API as much as possible, there are no destructor in Java, so the Interaction implements `java.io.Closeable`, and it will be up to the user to properly close the interaction.

Reviewed By: avalonalex

Differential Revision: D26627087

fbshipit-source-id: 48f57b0d38e87ccc0e6aa3a2e8c575bdad5055a5