Support for visiting a typedef's hierarchy to find annotations
Summary:
For example, the IDL could define a 3 level hierarchy:

  typedef i32 t1 (foo1 = "a");
  typedef t1 t2 (foo2 = "b");
  typedef t2 t3 (foo3 = "c");

then in the ast:

  t3->get_annotation("foo1")                                    // ""
  t3->get_annotation("foo2")                                    // ""
  t3->get_annotation("foo3")                                    // "c"

  t_typedef::get_first_annotation(t3, "foo1")                   // "a"
  t_typedef::get_first_annotation(t3, "foo2")                   // "b"
  t_typedef::get_first_annotation(t3, "foo3")                   // "c"

  t_typedef::get_first_annotation(t3, {"foo1", "foo2"})         // "b"
  t_typedef::get_first_annotation(t3, {"foo1", "foo2", "foo3"}) // "c"

I need this for the {lang.adapter} annotation. There is a lot of existing code tries to do this, but every impl I've seen is bugged, so I've made a correct version with test coverage we can switch too. I will also update the structured annotation logic to support the same methodology.

(I also made it so the default value isn't copied until the annotation is not found)

Reviewed By: vitaut

Differential Revision: D26602234

fbshipit-source-id: ab9ca38760149ce468dbc6851a0d7e651ee5cccb