static metadata interface
Summary:
Move to a staticmethod based metadata system so we don't have to create instances when passed classes.  Its faster to get the type of an instance than get an instance for a type.

Added type information to Proxy Classes, moved from public to readonly access to properties which is more inline with the rest of thrift-py3

Fixed enum comparisons not using 'is'

Fixed the type information to not use Sequence since they are Iterators and not a sequence container.

This opens up the ability to have ServiceInterfaces raise RuntimeErrors when not created inside a running loop.

Also instancing some structs can be expensive.

Reviewed By: Filip-F

Differential Revision: D26829306

fbshipit-source-id: 8b6016634e0563032fbf6bbfceb35a52c8f54bf2