Move cpp2 type resolution logic out of mustache
Summary:
The mustache logic was basically branching, which mustache is intentionally not designed for, so it was getting fairly complex and hard to update. A new c++ class replaces the mustache logic, also caches results.

I haven't actually changed any of the logic, so there are no changes to the fixtures.

Reviewed By: yfeldblum, vitaut

Differential Revision: D26800331

fbshipit-source-id: 26f8231a184f588cfb72a3bf38e894f8ca552f9e